requirements.txt
# WOPR Installer Dependencies
# ==========================

# Provider SDKs
hcloud>=1.30.0          # Hetzner Cloud (official SDK)
apache-libcloud>=3.8.0  # Multi-cloud abstraction (Vultr, DO, Linode, OVH)

# Billing & Payments
stripe>=7.0.0           # Stripe billing integration

# Web Framework (Dashboard API)
fastapi>=0.109.0        # Dashboard REST API
uvicorn>=0.27.0         # ASGI server
pydantic>=2.5.0         # Data validation
httpx>=0.26.0           # Async HTTP client (Authentik API)

# PDF Generation
reportlab>=4.0.0        # PDF document generation

# Database
asyncpg>=0.29.0         # Async PostgreSQL driver

# Email
jinja2>=3.1.0           # Email template rendering

# Utilities
paramiko>=3.0.0         # SSH for BYO provider
requests>=2.28.0        # HTTP client

# DNS automation
cloudflare>=2.11.0      # Cloudflare DNS API

# Security
slowapi>=0.1.0          # Rate limiting for FastAPI

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
